+++ CHANGELOG.md
@@ -2,6 +2,7 @@ QuantumAnnealingAnalytics.jl Change Log
 =======================================
 
 ### Staged
+- Change `plot_states` to show all states by default (fix)
 - Add ground energy shift option to hamiltonian energy plots
 - Add Ising model graph plotting